backend/requirements.txt
Django==2.1.7
pytz==2018.9
djangorestframework==3.9.2
django-filters==0.2.1
markdown==3.0.1
psycopg2==2.7.7
influxdb
requests